java/canaryeds-0.0/canaryeds-base/src/test/java/org/canaryeds/base/ConfigV4ConverterTest.java
/*
 * Copyright 2014 Sandia Corporation.
 * Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation, the U.S.
 * Government retains certain rights in this software.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * This software was written as part of an Inter-Agency Agreement between Sandia
 * National Laboratories and the US EPA NHSRC.
 */
package org.canaryeds.base;

import org.canaryeds.base.ConfigV4Converter;
import java.util.HashMap;
import org.junit.After;
import org.junit.AfterClass;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.fail;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Ignore;
import org.junit.Test;

@Ignore
public class ConfigV4ConverterTest {

    public ConfigV4ConverterTest() {
    }

    @BeforeClass
    public static void setUpClass() {
    }

    @AfterClass
    public static void tearDownClass() {
    }

    @Before
    public void setUp() {
    }

    @After
    public void tearDown() {
    }

    /**
     * Test of the ConfigV4Converter#convertV4toV5 method, of the ConfigV4Converter class.
     */
    @Test
    public void testConvertV4toV5() throws Exception {
        System.out.println("convertV4toV5");
        HashMap config = null;
        HashMap expResult = null;
        HashMap result = ConfigV4Converter.convertV4toV5(config);
        assertEquals(expResult, result);
        /// @todo review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

}
